Move repeated error message handling code into an include
