+++ client/src/components/StreamField/blocks/ListBlock.js
@@ -8,6 +8,10 @@ import {
 } from './BaseSequenceBlock';
 import { escapeHtml as h } from '../../../utils/text';
 import { range } from '../../../utils/range';
+import {
+  addErrorMessages,
+  removeErrorMessages,
+} from '../../../includes/streamFieldErrors';
 
 /* global $ */
 
@@ -256,19 +260,10 @@ export class ListBlock extends BaseSequenceBlock {
 
     // Non block errors
     const container = this.container[0];
-    container
-      .querySelectorAll(':scope > .help-block.help-critical')
-      .forEach((element) => element.remove());
+    removeErrorMessages(container);
 
     if (error.messages) {
-      // Add a help block for each error raised
-      error.messages.forEach((message) => {
-        const errorElement = document.createElement('p');
-        errorElement.classList.add('help-block');
-        errorElement.classList.add('help-critical');
-        errorElement.innerHTML = h(message);
-        container.insertBefore(errorElement, container.childNodes[0]);
-      });
+      addErrorMessages(container, error.messages);
     }
 
     if (error.blockErrors) {
