+++ client/src/components/StreamField/blocks/StreamBlock.js
@@ -18,6 +18,10 @@ import ComboBox, {
   comboBoxTriggerLabel,
 } from '../../ComboBox/ComboBox';
 import { hideTooltipOnEsc } from '../../../includes/initTooltips';
+import {
+  addErrorMessages,
+  removeErrorMessages,
+} from '../../../includes/streamFieldErrors';
 
 /* global $ */
 
@@ -401,19 +405,10 @@ export class StreamBlock extends BaseSequenceBlock {
 
     // Non block errors (messages applying to the block as a whole)
     const container = this.container[0];
-    container
-      .querySelectorAll(':scope > .help-block.help-critical')
-      .forEach((element) => element.remove());
+    removeErrorMessages(container);
 
     if (error.messages) {
-      // Add a help block for each error raised
-      error.messages.forEach((message) => {
-        const errorElement = document.createElement('p');
-        errorElement.classList.add('help-block');
-        errorElement.classList.add('help-critical');
-        errorElement.innerHTML = h(message);
-        container.insertBefore(errorElement, container.childNodes[0]);
-      });
+      addErrorMessages(container, error.messages);
     }
 
     if (error.blockErrors) {
